refactor track point interpolation out of component
