+++ src/components/Map.tsx
@@ -1,13 +1,13 @@
-import { Feature, Geometry, MultiPoint, Point, Polygon, Position } from "geojson";
+import { Feature, Geometry, Polygon } from "geojson";
 import { MapContainer,  GeoJSON } from 'react-leaflet'
 import { PathOptions, StyleFunction, LatLngExpression, CircleMarker, LeafletMouseEvent } from 'leaflet'
 import { TRACK_TYPE, ZONE_TYPE } from "../constants";
 import Track from "./Track";
 import Zone from "./Zone";
-import * as turf from "turf";
-import { useCallback, useEffect } from "react";
+import { useCallback, useEffect, useMemo } from "react";
 import { useAppSelector } from "../app/hooks";
 import { useAppContext } from "../context/AppContext";
+import { generateCurrentLocations } from "../helpers/generateCurrentLocations";
 
 interface CustomPathOptions extends PathOptions {
   radius?: number;
@@ -17,33 +17,10 @@ const isVisible = (feature: Feature): boolean => {
   return feature.properties?.visible
 }
 
-const isTemporal = (feature: Feature): boolean => {
-  return feature.properties?.times
-}
-
-const timeVal = (timeStr: string): number => {
-  return new Date(timeStr).getTime()
-}
-
-const calcInterpLocation = (poly: MultiPoint, times: any, current: number, index: number): Position => {
-  const coords = poly.coordinates
-  const isFirst = index === 0
-  const beforeIndex = isFirst ? 0 : index - 1
-  const afterIndex = isFirst ? 0 : index
-  const beforeCoords = coords[beforeIndex]
-  const afterCoords = coords[afterIndex]
-  const before = turf.point(beforeCoords)
-  const after = turf.point(afterCoords)
-  const turfPath = turf.lineString([beforeCoords, afterCoords])
-  const len = turf.distance(before, after)
-  const beforeTime = timeVal(times[beforeIndex])
-  const afterTime = timeVal(times[afterIndex])
-  const timeDelta = afterTime - beforeTime
-  const proportion = (current - beforeTime) / timeDelta
-  const lenProp = len * proportion
-  const interpolated = isNaN(lenProp) ? before : turf.along(turfPath, lenProp)
-  const markerLoc = interpolated.geometry.coordinates
-  return markerLoc
+const createLabelledPoint = (pointFeature: Feature, latlng: LatLngExpression, onTooltipClick: (event: LeafletMouseEvent) => void) => {
+  const color = pointFeature.properties?.color || 'blue';
+  const name = pointFeature.properties?.name || '';
+  return new CircleMarker(latlng, { radius: 1, fillOpacity: 1, color, opacity:1 }).bindTooltip(name, { interactive:true, permanent: true, direction: 'center' }).on('click', onTooltipClick);
 }
 
 interface MapProps {
@@ -54,7 +31,7 @@ const Map: React.FC<MapProps> = ({ children }) => {
   const features = useAppSelector(state => state.featureCollection.features)
   const { selection, setSelection, time, setCurrentLocations } = useAppContext();
 
-  const setColor: StyleFunction = (feature: Feature<Geometry, unknown> | undefined) => {
+  const styleFor: StyleFunction = (feature: Feature<Geometry, unknown> | undefined) => {
     const res: CustomPathOptions = {}
     if (feature) {
       const feat = feature as Feature
@@ -71,37 +48,12 @@ const Map: React.FC<MapProps> = ({ children }) => {
 
   useEffect(() => {
     if (time.current && features.length) {
-      const temporalFeatures = features.filter(isTemporal)
-      const pointFeatures = temporalFeatures.map((feature) => {
-        const times = feature.properties?.times
-        const timeNow = time.current
-        const index = times.findIndex((time: string) => new Date(time).getTime() >= timeNow)
-        if (index >= 0) {
-          const poly = feature.geometry as MultiPoint
-          const markerLoc = calcInterpLocation(poly, times, time.current, index)
-          const pointFeature: Feature<Point> = {
-            type: 'Feature',
-            id: feature.id,
-            geometry: {
-              type: 'Point',
-              coordinates: markerLoc
-            },
-            properties: {
-              name: feature.properties?.name,
-              color: feature.properties?.color
-            }
-          }
-          return pointFeature
-        }
-        return undefined
-      }).filter((f) => f !== undefined) as Feature<Point>[]
+      const pointFeatures = generateCurrentLocations(features, time)
       setCurrentLocations(pointFeatures)
     } else {
       setCurrentLocations([])
     }
   }, [time, features])
-
-
 
   const onClickHandler = useCallback((id: string, modifier: boolean): void => {
     if (modifier) {
@@ -123,29 +75,29 @@ const Map: React.FC<MapProps> = ({ children }) => {
     }
   }, [onClickHandler]);
 
-  const createLabelledPoint = (pointFeature: Feature, latlng: LatLngExpression) => {
-    const color = pointFeature.properties?.color || 'blue';
-    const name = pointFeature.properties?.name || '';
-    return new CircleMarker(latlng, { radius: 1, fillOpacity: 1, color, opacity:1 }).bindTooltip(name, { interactive:true, permanent: true, direction: 'center' }).on('click', onTooltipClick);
-  }
-  
   const featureFor = (feature: Feature): React.ReactElement => {
     switch(feature.properties?.dataType) {
     case TRACK_TYPE:
-      return <Track feature={feature} onClickHandler={onClickHandler}/> 
+      return <Track key={feature.id} feature={feature} onClickHandler={onClickHandler}/> 
     case ZONE_TYPE:
-      return <Zone feature={feature as Feature<Polygon>} onClickHandler={onClickHandler}/>  
+      return <Zone key={feature.id} feature={feature as Feature<Polygon>} onClickHandler={onClickHandler}/>  
     default:
-      return <GeoJSON key={`${feature.id || 'index'}`} data={feature} style={setColor} pointToLayer={createLabelledPoint} /> 
+      return <GeoJSON key={`${feature.id || 'index'}`} data={feature} style={styleFor} 
+      pointToLayer={(pointFeature, latlng) => createLabelledPoint(pointFeature, latlng, onTooltipClick)} /> 
     }
   }
+
+  const visibleFeatures = useMemo(() => {
+    const vis = features.filter(feature => isVisible(feature))
+    return vis.map(featureFor)
+  }, [features])
 
   return (
     <>
       <MapContainer center={[35.505, -4.09]} zoom={8} scrollWheelZoom={true}>
         {children}
         { 
-          features.filter(feature => isVisible(feature)).map((featureFor))
+          visibleFeatures
         }
       </MapContainer>
     </>
